Remove numeric framework shim
This migrates the remaining core API tests to call the new
implementation directly and removes all of the remaining shim code
that existed to translate between the new and old APIs.

Issue #2416
