+++ src/unittests/serialization.spec.ts
@@ -31,8 +31,7 @@ import {
   vec3,
   vec4,
 } from '../webgpu/util/conversion.js';
-import { toF32Interval } from '../webgpu/util/f32_interval.js';
-import { deserializeFPInterval, serializeFPInterval } from '../webgpu/util/floating_point.js';
+import { deserializeFPInterval, FP, serializeFPInterval } from '../webgpu/util/floating_point.js';
 
 import { UnitTest } from './unit_test.js';
 
@@ -218,28 +217,28 @@ g.test('value').fn(t => {
 
 g.test('f32_interval').fn(t => {
   for (const interval of [
-    toF32Interval(0),
-    toF32Interval(-0),
-    toF32Interval(1),
-    toF32Interval(-1),
-    toF32Interval(0.5),
-    toF32Interval(-0.5),
-    toF32Interval(kValue.f32.positive.max),
-    toF32Interval(kValue.f32.positive.min),
-    toF32Interval(kValue.f32.subnormal.positive.max),
-    toF32Interval(kValue.f32.subnormal.positive.min),
-    toF32Interval(kValue.f32.subnormal.negative.max),
-    toF32Interval(kValue.f32.subnormal.negative.min),
-    toF32Interval(kValue.f32.infinity.positive),
-    toF32Interval(kValue.f32.infinity.negative),
+    FP.f32.toInterval(0),
+    FP.f32.toInterval(-0),
+    FP.f32.toInterval(1),
+    FP.f32.toInterval(-1),
+    FP.f32.toInterval(0.5),
+    FP.f32.toInterval(-0.5),
+    FP.f32.toInterval(kValue.f32.positive.max),
+    FP.f32.toInterval(kValue.f32.positive.min),
+    FP.f32.toInterval(kValue.f32.subnormal.positive.max),
+    FP.f32.toInterval(kValue.f32.subnormal.positive.min),
+    FP.f32.toInterval(kValue.f32.subnormal.negative.max),
+    FP.f32.toInterval(kValue.f32.subnormal.negative.min),
+    FP.f32.toInterval(kValue.f32.infinity.positive),
+    FP.f32.toInterval(kValue.f32.infinity.negative),
 
-    toF32Interval([-0, 0]),
-    toF32Interval([-1, 1]),
-    toF32Interval([-0.5, 0.5]),
-    toF32Interval([kValue.f32.positive.min, kValue.f32.positive.max]),
-    toF32Interval([kValue.f32.subnormal.positive.min, kValue.f32.subnormal.positive.max]),
-    toF32Interval([kValue.f32.subnormal.negative.min, kValue.f32.subnormal.negative.max]),
-    toF32Interval([kValue.f32.infinity.negative, kValue.f32.infinity.positive]),
+    FP.f32.toInterval([-0, 0]),
+    FP.f32.toInterval([-1, 1]),
+    FP.f32.toInterval([-0.5, 0.5]),
+    FP.f32.toInterval([kValue.f32.positive.min, kValue.f32.positive.max]),
+    FP.f32.toInterval([kValue.f32.subnormal.positive.min, kValue.f32.subnormal.positive.max]),
+    FP.f32.toInterval([kValue.f32.subnormal.negative.min, kValue.f32.subnormal.negative.max]),
+    FP.f32.toInterval([kValue.f32.infinity.negative, kValue.f32.infinity.positive]),
   ]) {
     const serialized = serializeFPInterval(interval);
     const deserialized = deserializeFPInterval(serialized);
@@ -256,10 +255,10 @@ g.test('expression_expectation').fn(t => {
     f32(123),
     vec2(f32(1), f32(2)),
     // Interval
-    toF32Interval([-0.5, 0.5]),
-    toF32Interval([kValue.f32.positive.min, kValue.f32.positive.max]),
+    FP.f32.toInterval([-0.5, 0.5]),
+    FP.f32.toInterval([kValue.f32.positive.min, kValue.f32.positive.max]),
     // Intervals
-    [toF32Interval([-8.0, 0.5]), toF32Interval([2.0, 4.0])],
+    [FP.f32.toInterval([-8.0, 0.5]), FP.f32.toInterval([2.0, 4.0])],
   ]) {
     const serialized = serializeExpectation(expectation);
     const deserialized = deserializeExpectation(serialized);
