--- src/webgpu/util/f32_interval.ts
@@ -1,37 +1,0 @@
-// This is a shim file with all the old/deprecated f32 API calls.
-// They currently just pass-through to the refactored FPContext implementation.
-// As CTS migrates over to directly calling the new API, these will be removed.
-
-import { FPInterval, FPMatrix, FPVector, IntervalBounds, FP } from './floating_point.js';
-
-// Containers
-
-export type F32Vector = FPVector;
-
-// Utilities
-
-export function toF32Interval(n: number | IntervalBounds | FPInterval): FPInterval {
-  return FP.f32.toInterval(n);
-}
-
-export function isF32Vector(v: (number | IntervalBounds | FPInterval)[]): v is FPVector {
-  return FP.f32.isVector(v);
-}
-
-export function toF32Vector(v: (number | IntervalBounds | FPInterval)[]): FPVector {
-  return FP.f32.toVector(v);
-}
-
-export function spanF32Intervals(...intervals: FPInterval[]): FPInterval {
-  return FP.f32.spanIntervals(...intervals);
-}
-
-export function isF32Matrix(
-  m: (number | IntervalBounds | FPInterval)[][] | FPVector[]
-): m is FPMatrix {
-  return FP.f32.isMatrix(m);
-}
-
-export function toF32Matrix(m: (number | IntervalBounds | FPInterval)[][] | FPVector[]): FPMatrix {
-  return FP.f32.toMatrix(m);
-}
